+++ rnalysis/gui/gui.py
@@ -1955,7 +1955,7 @@ class FilterTabPage(TabPage):
                         'split_hierarchical': 'Hierarchical (Agglomerative)', 'split_hdbscan': 'HDBSCAN',
                         'split_clicom': 'CLICOM (Ensemble)'}
     SUMMARY_FUNCS = {'describe', 'head', 'tail', 'biotypes_from_ref_table', 'biotypes_from_gtf', 'print_features'}
-    GENERAL_FUNCS = {'sort', 'sort_by_principal_component', 'transform', 'translate_gene_ids',
+    GENERAL_FUNCS = {'concatenate', 'sort', 'sort_by_principal_component', 'transform', 'translate_gene_ids',
                      'differential_expression_deseq2', 'differential_expression_deseq2_simplified', 'fold_change',
                      'average_replicate_samples', 'drop_columns',
                      'differential_expression_limma_voom', 'differential_expression_limma_voom_simplified',