+++ HISTORY.rst
@@ -12,6 +12,7 @@ Added
 * bowtie2 and kallisto paired-end modes now support the new file naming method 'smart', that attempts to automatically determine the common name of each pair of paired-end fastq files.
 * Added a p-value histogram plot (DESeqFilter.pval_histogram) that displays the distribution of p-values in a differential expression table.
 * Added new visualization options for 'plot expression of specific genes' function (CountFilter.plot_expression).
+* Added new function 'Concatenate two tables' (Filter.concatenate) that concatenates two tables along their rows.
 
 Changed
 ********